refactoring a bit
